redone database implementation

+++ app/src/main/java/com/example/weatherapp/entities/Wind.kt
@@ -1,3 +1,3 @@
 package com.example.weatherapp.entities
 
-data class Wind(val speed: Double, val deg: Double)
+data class Wind(val speed: Float, val deg: Float)
